+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/compiler/visitors/outer/CircuitRewriter.java
@@ -367,6 +367,6 @@ public class CircuitRewriter extends CircuitCloneVisitor {
 
     @Override
     public String toString() {
-        return super.toString() + ":" + this.transform;
+        return super.toString() + "-" + this.transform;
     }
 }